Create trackers after onboarding and simplified tracker creation code
